libpq: add OAUTHBEARER SASL mechanism
DO NOT USE THIS PROOF OF CONCEPT IN PRODUCTION.

Implement OAUTHBEARER (RFC 7628) and OAuth 2.0 Device Authorization
Grants (RFC 8628) on the client side. When speaking to a OAuth-enabled
server, it looks a bit like this:

    $ psql 'host=example.org oauth_client_id=f02c6361-0635-...'
    Visit https://oauth.example.org/login and enter the code: FPQ2-M4BG

The OAuth issuer must support device authorization. No other OAuth flows
are currently implemented (but clients may provide their own flows; see
below).

The client implementation requires either libcurl or libiddawc and their
development headers. Pass `curl` or `iddawc` to --with-oauth/-Doauth
during configuration.

Thomas Munro wrote the kqueue() implementation for oauth-curl; thanks!

= PQauthDataHook =

Clients may override two pieces of OAuth handling using the new
PQsetAuthDataHook():

- PQAUTHDATA_PROMPT_OAUTH_DEVICE: replaces the default user prompt to
  standard error when using the builtin device authorization flow

- PQAUTHDATA_OAUTH_BEARER_TOKEN: replaces the entire OAuth flow with a
  custom asynchronous implementation

In general, a hook implementation should examine the incoming `type` to
decide whether or not to handle a specific piece of authdata; if not, it
should delegate to the previous hook in the chain (retrievable via
PQgetAuthDataHook()). Otherwise, it should return an integer > 0 and
follow the authdata-specific instructions. Returning an integer < 0
signals an error condition and abandons the connection attempt.

== PQAUTHDATA_PROMPT_OAUTH_DEVICE ==

The hook should display the device prompt (URL + code) using whatever
method it prefers.

== PQAUTHDATA_OAUTH_BEARER_TOKEN ==

The hook should either directly return a Bearer token for the current
user/issuer/scope combination, if one is available without blocking, or
else set up an asynchronous callback to retrieve one. See the
documentation for PQoauthBearerRequest.

Several TODOs:
- don't retry forever if the server won't accept our token
- perform several sanity checks on the OAuth issuer's responses
- handle cases where the client has been set up with an issuer and
  scope, but the Postgres server wants to use something different
- improve error debuggability during the OAuth handshake
- fix libcurl initialization thread-safety
- harden the libcurl flow implementation
- figure out pgsocket/int difference on Windows
- ...and more.